requirements.txt
python-dotenv
huggingface_hub
google-cloud-aiplatform
vertexai
openai
pandas
matplotlib